pyproject.toml
[project]
name = "langchain-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langgraph-checkpoint-mysql[pymysql]>=3.0.0",
]
